requirements.txt
Flask
scrapetube
pytube
google-generativeai
python-dotenv
youtube-transcript-api
requests
beautifulsoup4
